requirements.txt
asgiref==3.8.1
attrs==25.3.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
coreschema==0.0.4
cryptography==45.0.2
decorator==5.2.1
Django==5.2.1
django-cors-headers==4.7.0
django-filter==25.1
django-oauth-toolkit==3.0.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
et_xmlfile==2.0.0
idna==3.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.6
joblib==1.5.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jwcrypto==1.5.6
llvmlite==0.44.0
MarkupSafe==3.0.2
msgpack==1.1.0
mysqlclient==2.2.7
numpy==2.2.6
oauthlib==3.2.2
openpyxl==3.1.5
packaging==25.0
pillow==11.2.1
platformdirs==4.3.8
pycparser==2.22
Pygments==2.19.1
PyJWT==2.9.0
python-dotenv==1.1.0
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.25.0
scipy==1.15.3
soundfile==0.13.1
soxr==0.5.0.post1
sqlparse==0.5.3
standard-aifc==3.13.0
standard-chunk==3.13.0
standard-sunau==3.13.0
threadpoolctl==3.6.0
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.4.0
